+++ src/pages/_app.tsx
@@ -1,10 +1,9 @@
-import '@/styles/globals.css';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { ModalProvider } from '@/shared/providers/modal-provider';
-import type { AppProps } from 'next/app';
 import { CookiesProvider } from 'react-cookie';
-import useRouterGuard from '@/shared/hooks/useRouterGuard';
+import type { AppProps } from 'next/app';
+import { ModalProvider } from '@/shared/providers/modal-provider';
+import '@/styles/globals.css';
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -16,8 +15,6 @@ const queryClient = new QueryClient({
 });
 
 export default function App({ Component, pageProps }: AppProps) {
-  useRouterGuard();
-
   return (
     <QueryClientProvider client={queryClient}>
       <CookiesProvider defaultSetOptions={{ path: '/' }}>
